JSONLines

A simple package to read (parts of) a JSON Lines files to a vector of JSON3.Objects. This vector is Tables.jl compatible and can directly be piped into e.g. a DataFrame if every row in the result has the same schema (i.e. the same variables). It allows memory-efficient loading of rows of a JSON Lines file. In order to select the rows skip and nrows can be used to load nrows rows after skipping skip rows. The file is mmaped and only the required rows are loaded into RAM. Files must contain a valid JSON object (denoted by {"String1":ELEMENT1, "String2":ELEMENT2, ...}) on each line. JSON parsing is done using the JSON3.jl package. Lines can be separated by \n or \r\n and some whitespace characters are allowed between the end of the JSON object and the newline character (basically all that can be represented as a single UInt8). Typically a file would look like this:

{"name":"Daniel","organization":"IMSM"}
{"name":"Peter","organization":"StatMath"}

Getting Started

(@v1.5) pkg> add JSONLines

Functions

JSONLines.readfile

JSONLines.readlazy

JSONLines.reset!

JSONLines.writefile

JSONLines.readfile — Method

readfile(file::AbstractString; kwargs...) => Vector{JSON3.Object}

Read (parts of) a JSONLines file.

file: Path to JSONLines file

Keyword Arguments:

structtype = nothing: StructType passed to JSON3.read for each row of the file

nrows = nothing: Number of rows to load

skip = nothing: Number of rows to skip before loading

usemmap::Bool = (nrows !== nothing || skip !=nothing): Memory map file (required for nrows and skip)

nworkers::Int = 1: Number of threads to spawn for parsing the file

JSONLines.readlazy — Method

readlazy(file; structtype = nothing, skip = nothing) => JSONLines.LazyRows

Get a lazy iterator over a JSONLines file. iterate(l::LazyRows) returns a Tuple with the JSON3.Object of the current line and the index of its last element. A LazyRows object tracks its own state (which can be reset to the beginning of the file using reset!) so that it is possible to call iterate without additional arguments. To materialize all elements call [row for row in readlazy("file.jsonl")].

file: Path to JSONLines file

structtype = nothing: StructType passed to JSON3.read for each row of the file

skip = nothing: Number of rows to skip at the beginning of the file

JSONLines.reset! — Method

reset!(rows::LazyRows)

Reset row iterator to beginning of file

JSONLines.writefile — Function

writefile(file, data, mode = "w")

Write data to file in the JSONLines format.

file: Path to target JSONLines file

data: Tables.jl compatible data source

mode = "w": Mode to open the file in see I/O and Network